Rename `httputil` to `httpx`.
